ci: create a workflow to test the commit lint action (#7)
* ci: create a test for the commit lint action

Test that a single good commit is correctly passed.
Test that a single bad commit is failed.
Test that multiple bad commits are failed.

* ci: use --from option to lint since the last tag

This fixes an issue where the oldest commit was not being checked because the "while read" structure wasn't executing for the final line of output from git log.